# openapi.yaml
openapi: 3.1.0
info:
  title: Tooxs BUK API
  version: 1.0.0
servers:
  - url: https://tooxs-buk-api.onrender.com
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Verifica que la API este activa
      responses:
        "200":
          description: Estado del servicio
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  mode:
                    type: string
  /chat:
    post:
      operationId: consultarBUK
      summary: Consulta empleados y planilla de Tooxs en lenguaje natural
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Respuesta del agente
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
      security:
        - ApiKeyAuth: []
  /payroll/upload-json:
    post:
      operationId: uploadPayrollJson
      summary: Carga manual de planilla real por periodo (JSON)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - period
                    - total_clp
                  properties:
                    period:
                      type: string
                      example: 2026-03
                    total_clp:
                      type: number
                    currency:
                      type: string
                      example: CLP
                    employees_count:
                      type: integer
                - type: array
                  items:
                    type: object
      responses:
        "201":
          description: Registros guardados
      security:
        - ApiKeyAuth: []
  /payroll/upload-csv:
    post:
      operationId: uploadPayrollCsv
      summary: Carga manual de planilla real por periodo (CSV)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Registros guardados
      security:
        - ApiKeyAuth: []
  /payroll/periods:
    get:
      operationId: getPayrollPeriods
      summary: Lista periodos de planilla almacenados
      responses:
        "200":
          description: Lista de periodos
      security:
        - ApiKeyAuth: []
  /payroll/{period}:
    get:
      operationId: getPayrollByPeriod
      summary: Obtiene planilla real de un periodo
      parameters:
        - in: path
          name: period
          required: true
          schema:
            type: string
            example: 2026-03
      responses:
        "200":
          description: Planilla del periodo
        "404":
          description: No existe planilla para ese periodo
      security:
        - ApiKeyAuth: []
  /webhooks/buk/payroll:
    post:
      operationId: receiveBukPayrollWebhook
      summary: Recibe eventos de cierre de planilla desde BUK
      description: Requiere header X-Webhook-Secret con el valor de BUK_WEBHOOK_SECRET
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Evento recibido